+++ src/main/java/com/alma/finantrack/models/dao/ICategoriaDAO.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import com.alma.finantrack.models.entity.Categoria;
 
 @Repository
-public interface ICategoriaDAO extends JpaRepository<Categoria, Long> {
+public interface ICategoriaDAO extends JpaRepository<Categoria, Integer> {
 }
