+++ src/main/java/com/alma/finantrack/models/dao/IPresupuestoDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPresupuestoDAO extends CrudRepository<Presupuesto, Long> {
-    // Puedes agregar métodos personalizados si es necesario
+public interface IPresupuestoDAO extends CrudRepository<Presupuesto, Integer> {
+
 }
