+++ src/main/java/com/alma/finantrack/models/dao/ITransaccionDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITransaccionDAO extends JpaRepository<Transaccion, Long> {
+public interface ITransaccionDAO extends JpaRepository<Transaccion, Integer> {
     // Aquí puedes agregar métodos personalizados si es necesario
 }
